examples/client/app/requirements.txt
httpx
openai
python-dotenv